Change statement order to highlight duplication
In the course of reading the code around ModelSet<T> for making
DependencySpecContainer a proper subtype of it I've found a huge
chunk of duplicated logic between
`ManagedSetNodeInitializerExtractionStrategy` and
`ModelSetNodeInitializerExtractionStrategy`.

This commit doesn't solve it but makes it easier to spot it and
extract it in the future by reordering statements to be the same in both
places.