Don't build the runtime jar for a resolved library
By making the `ApiJar` a first class, buildable concept in the Java Software Model,
this change takes a small step toward decomposing `JarBinarySpec` into a
"graph of buildable things".

The `ApiJar` is not yet part of the public API, but is used when resolving
a library dependency, providing the correct api artifact as well as wiring in
the required tasks.