Added some thread safety to `ResolutionResultsStoreFactory`
Hopefully this should fix GRADLE-2289, as well as some of the recent
sporadic failures around IDE project generation. The root cause was that
the incorrect resolution result could be streamed from disk when resolving
in parallel.

We all know that '++' isn't atomic. Fun times when it's causing the results
of one configuration to be replaced with another.