Fix integration tests: we don't configure tasks for empty Twirl and Routes source sets
The language infrastructure avoids creating tasks for source sets that don't have any
sources. This was done because we _create_ a source set for every applicable language,
and we wanted to avoid the user see a compile task for every one of these when their
component only used one type of source.

So if a project applies a bunch of language plugins, we wanted to only create
the CompileCpp task if there were actually Cpp sources present. 
A better approach to achieve this might be to avoid creating these empty source sets 
in the first place.