Changed model report so that rule binding validations are applied prior to generating the report.
This means that the model report can be used as a way to exercise and validate as many rules as possible, eg by build authors.

The implementation isn't quite right. This validation should really happen as part of closing an element. On making an element immutable, we should fail early when there are rules whose references to the element or its descendants can never be satisfied.

+review REVIEW-5756