ManagedModelProjection produce correct value description when no value
In the managed world, a "no value" node is-a `ModelReferenceNode` with a
`null` `target`.
Previous code created a `ModelView` over the node that assumed it had
some value in it, always.
This produced wrong reporting for "no value" nodes, hence the test
changes included.
This commit fixes just that by asking the node to directly create a
proper `ModelView` so that the "no value" case is handled directly by
`ModelReferenceNode` itself.

This should fix the IBM JDK 6 VerifyError along the way as the generated
`ModelView` is now aware of its state.

+review REVIEW-5734