Changed the implementation for binding by-type model references so that it is much more efficient.
In particular, changed `RuleBindings` to reuse the result of a (scope, type) match for all later references to that (scope, type).

Also separately handle untyped by-path references and  by-path + by-type references to reduce the amount of work done when a node becomes known and when a node's type information becomes available.

Added more unit test coverage for `RuleBindings`.