Support for testing a JVM component without dependencies
This commit provides a first working implementation of a component under test for the Java software model. Some adjustments were necessary to be able to tweak the compile classpath of the test suite without having to attach the sources and dependencies of the tested component to the test suite itself.

Story: gradle/langos#113