Address Review comments for Gradle-1945
The configuration name is no longer logged, as it was not the name of the configuration
where the dependency was originally declared. This would only have confused users.
The logging behavior of Idea was changed to match that of Eclipse. Dependencies
which are filtered out using scopes.minus will no longer be logged.

+review REVIEW-5782