Initial support for declaring JDKs in the model
This commit is the initial introduction of declaration of JDKs in the Java software model. It allows a user to declare a list of JDKs, and consequently a list of `InstalledJdk` is populated with the result of probing the versions from the executable. This commit does *not* include any error handling, nor does it try to be smart in displaying the type of JDK (for example, inferring that OpenJDK is used).

Story: gradle/langos#153