Add 'compileOnly' configuration for each source set.
This commit creates a new 'compileOnly' configuration for each declared source set. Dependencies added to this source set are used during compilation only. They are not included on the runtime classpath, are not inherited by test classpath, are not included in 'deployable' artifacts such as WARs, EARs or application distributions, are not exported to consuming projects, and are not included in published metadata.

+review REVIEW-5807