Rename Buildable{ModelElement => ComponentSpec}
This rename better reflects the fact that (the former)
BuildableModelElement now extends ComponentSpec, and in the process
helps to eliminate the unnecessary concept of "model elements" in favor
of referring to everything in terms of components and component specs.

Javadoc in BuildableComponentSpec and certain parameters and member
variables in AbstractBuildableComponentSpec have also been updated and
renamed for clarity and consistency.