In the daemon composite coordinator, load and execute participant builds in process
In order to make it easier to create an "integrated" composite build, participant
builds are now loaded in process via `GradleLauncher`. This will permit shared services
to be injected into these builds, to allow for automatic dependency substitution.

A later implementation may allow these builds to be executed in separate daemon
processes, with remote communication and integration via Tooling API.