Make sure we use TAPI classes from the version under test
This commit introduces a `tapiClass` method to be used whenever we reference a class that belongs to the tooling API.
The reason is that when the action closure is created, we are using the TAPI classes of the test classpath, which is
the current version of the TAPI. But when the test is executed, we want to use the version under test. Ideally we
should have a distinct test runner to deal with this.