Don't alias composite project dependencies in generated IDEA modules
We now assume that all project dependency names are unique in a composite.
This is a temporary hack to make it possible to try out generating the IDEA
metadata files for a composite.

Dependencies are correctly substituted in generated module files,
but project file does not include modules for substituted project.