Add link icon to section headers.
This feature makes it easy for users to get the direct URL for a given section
of the user guide. Hovering over a section header makes a link icon appear next
to it. That link icon contains a direct link to the corresponding section.

The implementation is almost pure CSS. The anchor element has been given an SVG
as its background image, hidden by default. The image and link appear via a
`:hover` rule on the header element. I did need to customise the HTML generated
from the DocBook so that the anchors had `href` attributes. This was done by
copying the existing XSL rule and making the necessary adjustment to it.