Added `EclipseProjectDependency.targetProjectDirectory`
This removes the need to always have an `EclipseProject` instance to target in
a dependency, which is the case for inter-build dependencies in a composite.
Buildship is actually doing this transformation (targetProject -> targetProjectDirectory),
so it makes sense to put it into the Eclipse model.

This may be enhanced further to provide a true identifier for each `EclipseProject`,
and reference by the identifier.

A compatibility mapper makes this value available for all target Gradle versions >= 1.0