Improve the error message when accessing an unknown property of a decorated object from the DSL.
In particular, the error message is now much more likely to report the actual object that the property access was performed on. For example, instead of 'unknown property foo on task set' you will get 'unknown property foo on project :core', or instead of 'unknown property foo on type Convention' you will get 'unknown property foo on task :compileJava'.

There are still some edge cases, and more context could be included in the error message, but at least the  message is far less likely to be misleading now.